pyproject.toml
[build-system]
requires = ["uv_build >= 0.9.21, <0.10.0"]
build-backend = "uv_build"

[project]
name = "slapyshot"
version = "0.0.1"
description = "A Python wrapper for the SportRadar NHL API"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.32.0",
    "python-dotenv>=1.0.0",
    "polars>=1.0.0",
]
authors = [
  { name = "Caleb Peterson", email = "calebpeterson22@gmail.com" }
]

maintainers = [
  { name = "Caleb Peterson", email = "calebpeterson22@gmail.com" }
]

[project.optional-dependencies]
dev = [
    "pytest",
    "ruff",
]

[dependency-groups]
dev = [
    "pytest",
    "ruff",
]